card game runs but sucks

+++ testing/src/Board.java
@@ -117,7 +117,7 @@ public abstract class Board {
 	 */
 	public void replaceSelectedCards(List<Integer> selectedCards) {
 		for (Integer k : selectedCards) {
-			deal(k.intValue());
+			deal(k);
 		}
 	}
